+++ NetduinoApplicationPololu/Matrix.cs
@@ -37,9 +37,9 @@ with MinIMU-9-Arduino-AHRS. If not, see <http://www.gnu.org/licenses/>.
 
 /**************************************************/
 //Multiply two 3x3 matrixs. This function developed by Jordi can be easily adapted to multiple n*n matrix's. (Pero me da flojera!). 
-public static void Matrix_Multiply(float[][] a, float[][] b,float[][] mat)
+public static void Matrix_Multiply(double[][] a, double[][] b,double[][] mat)
 {
-  float[] op = new float[3]; 
+  double[] op = new double[3]; 
   for(int x=0; x<3; x++)
   {
     for(int y=0; y<3; y++)
@@ -51,7 +51,7 @@ public static void Matrix_Multiply(float[][] a, float[][] b,float[][] mat)
       mat[x][y]=0;
       mat[x][y]=op[0]+op[1]+op[2];
       
-      float test=mat[x][y];
+      double test=mat[x][y];
     }
   }
 }
